Feat: fixed library embla carousel
